Sample layout, fluid, slider,colors
Added a simple sample layout showing how to use the framework.

Implemented fluid container classes.

Reimplemented the slider. The functionality is the same but it is
much more robust and can handle more than one slider now.

Added metro-colors.css, these are the standard colors for Windowsphone
which are suited for a metro layout