Implemented image and header classes
Added header classes that are half a tile in height, but there should also
be som that are a quarter (normal header).

Renamed tile-container-* to metro-container-* to show that it can contain
any kind of metro element.

Implemented image background classes that rely on a /images folder that
doesn't ship yet.

Changed index.html to reflect these changes and to feel more alive (metro).